+++ include/posix/time.h
@@ -81,7 +81,8 @@ extern "C" {
 
 static inline int32_t _ts_to_ms(const struct timespec *to)
 {
-	return (to->tv_sec * MSEC_PER_SEC) + (to->tv_nsec / NSEC_PER_MSEC);
+	return (int32_t)((to->tv_sec * (int32_t)MSEC_PER_SEC) +
+		(to->tv_nsec / (int32_t)NSEC_PER_MSEC));
 }
 
 #ifdef CONFIG_ARCH_POSIX
